requirements.txt
streamlit
opencv-python-headless==4.8.0.76
pandas
numpy
pytesseract
Pillow
matplotlib